--- test/tokens/v3/clankerV3Token.validate.test.ts
@@ -1,3 +1,0 @@
-import { describe } from 'bun:test';
-
-describe('validations', async () => {});
